requirements.txt.txt
Flask
openai
pdf2image
pytesseract
Pillow
